Issue #7 Updated the get and update APIs for WorryHeads
